Added A Construction Filter
You can now only place blocks on certain other blocks, for example you cannot place a floor block on a wall block, but you can put a wall block on a floor block. This code is going to get very bloated very fast, but that's a problem for future me to deal with.

+++ Assets/Scripts/ObjectData.cs
@@ -7,5 +7,5 @@ public class ObjectData : MonoBehaviour
     public string idName;
     public int id;
 
-    public float x, y;
+    public int x, y;
 }
